Overhaul code to allow asynchronous tweet collection
Allow async tweet collection to finalize the Pyckaxe class. It now does
not need to rely on keyboard interrupts to exit. To do this I also had
to thread the databasehandler as sqlite3 databases cannot be used in
threads in which they are not created (running async runs collection in
a thread).

So basically, we needed a DIFFERENT thread to create and manage the
database.